requirements.txt
aiogram==3.0.0b7
fastapi==0.101.0
loguru==0.7.0
python-dotenv==1.0.0
SQLAlchemy==2.0.20
uvicorn==0.23.2

openpyxl~=3.1.2
alembic~=1.11.3
aiofiles==23.1.0
pydantic==1.10.10
pandas==2.2.2

python-jose~=3.4.0
passlib~=1.7.4